+++ tasks/generate-rule-exports.ts
@@ -4,6 +4,8 @@ import path from 'node:path';
 
 import fg from 'fast-glob';
 
+import { formatJson } from '../src/utils/format-json';
+
 // eslint-disable-next-line @typescript-eslint/naming-convention, no-underscore-dangle
 const __dirname = new URL('.', import.meta.url).pathname;
 
@@ -14,26 +16,28 @@ const packageJsonPath = path.join(__dirname, '../package.json');
  */
 async function main() {
     const files = await fg([path.join(__dirname, '../src/rules/*.ts')]);
-    const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf-8'));
+    const originalRawPkg = await readFile(packageJsonPath, 'utf-8');
+    const pkg = JSON.parse(originalRawPkg);
 
     // clear existing rule exports
-    for (const key of Object.keys(packageJson.exports)) {
+    for (const key of Object.keys(pkg.exports)) {
         if (key.startsWith('./rules/')) {
-            delete packageJson.exports[key];
+            delete pkg.exports[key];
         }
     }
 
     for (const file of files) {
         const { name } = path.parse(file);
 
-        packageJson.exports[`./rules/${name}`] = {
+        pkg.exports[`./rules/${name}`] = {
             types: `./dist/rules/${name}.d.ts`,
             import: `./dist/rules/${name}.js`,
             default: `./dist/rules/${name}.js`,
         };
     }
 
-    await writeFile(packageJsonPath, `${JSON.stringify(packageJson, null, 4)}\n`);
+    const updatedRawPkg = formatJson(pkg, originalRawPkg);
+    await writeFile(packageJsonPath, updatedRawPkg);
 }
 
 main();